refactor kurtosis feature extractor
