--- laserchicken/feature_extractor/kurtosis_norm_z_feature_extractor.py
@@ -1,11 +1,0 @@
-from laserchicken import keys
-from laserchicken.feature_extractor.kurtosis_z_feature_extractor import KurtosisZFeatureExtractor
-
-
-class KurtosisNormZFeatureExtractor(KurtosisZFeatureExtractor):
-    """Calculates the variation on the normalized height."""
-    DATA_KEY = keys.normalized_height
-
-    @classmethod
-    def provides(cls):
-        return ['kurto_norm_z']
