+++ laserchicken/feature_extractor/test_kurtosis_z_feature_extractor.py
@@ -2,8 +2,9 @@ import unittest
 
 import numpy as np
 
-from laserchicken import read_las
-from laserchicken.feature_extractor.kurtosis_z_feature_extractor import KurtosisZFeatureExtractor
+from laserchicken import read_las, keys
+from laserchicken.feature_extractor.kurtosis_feature_extractor import KurtosisFeatureExtractor
+from laserchicken.test_tools import create_point_cloud
 
 
 class TestKurtosisZFeatureExtractor(unittest.TestCase):
@@ -12,11 +13,38 @@ class TestKurtosisZFeatureExtractor(unittest.TestCase):
         neighborhood = [89664, 23893, 30638, 128795, 62052, 174453, 29129, 17127, 128215, 29667, 116156, 119157, 98591,
                         7018,
                         61494, 65194, 117931, 62971, 10474, 90322]
-        kurtosis_z = KurtosisZFeatureExtractor().extract(pc_in, neighborhood, None, None, None)
+        kurtosis_z = self.extractor.extract(pc_in, neighborhood, None, None, None)
         np.testing.assert_allclose(kurtosis_z, 3.968414258629714)
 
     def test_height_stats_without_neighbors(self):
         pc_in = read_las.read("testdata/AHN2.las")
         neighborhood = []
-        kurtosis_z = KurtosisZFeatureExtractor().extract(pc_in, neighborhood, pc_in, None, None)
+        kurtosis_z = self.extractor.extract(pc_in, neighborhood, pc_in, None, None)
         assert np.isnan(kurtosis_z)
+
+    def test_default_provides_correct(self):
+        feature_names = self.extractor.provides()
+        self.assertIn('kurto_z', feature_names)
+
+    def setUp(self):
+        self.extractor = KurtosisFeatureExtractor()
+
+
+class TestKurtosisNormZFeatureExtractor(unittest.TestCase):
+    def test_use_norm_z(self):
+        x = y = np.array([0, 0, 0])
+        z = np.array([2, 2, 2])
+        normalized_z = np.array([3, 4, 5])
+        point_cloud = create_point_cloud(x, y, z, normalized_z=normalized_z)
+        neighborhood = [[0, 1, 2]]
+
+        kurtosis = self.extractor.extract(point_cloud, neighborhood, None, None, None)
+
+        self.assertAlmostEqual(kurtosis, -1.5)
+
+    def test_default_provides_correct(self):
+        feature_names = self.extractor.provides()
+        self.assertIn('kurto_normalized_height', feature_names)
+
+    def setUp(self):
+        self.extractor = KurtosisFeatureExtractor(data_key=keys.normalized_height)
